Seperate utility functions from components
